+++ frontend/src/components/Header/SearchBar/SelectItemArea/ButtonArea/ButtonArea.tsx
@@ -1,9 +1,8 @@
 import CloseIcon from "@mui/icons-material/Close";
 import SearchIcon from "@mui/icons-material/Search";
-import { IconButtonProps } from "@mui/material";
 
 import { SelectItemTemplate } from "../SelectItemTemplate/SelectItemTemplate";
-import RoundButton from "./ButtonArea.style";
+import RoundButton, { RoundButtonProps } from "./ButtonArea.style";
 
 const icons = {
   close: <CloseIcon />,
@@ -45,12 +44,7 @@ const ButtonArea = ({
 
 export default ButtonArea;
 
-export interface SearchBarButtonProps extends IconButtonProps {
-  icon: "close" | "search";
-  isFocused?: boolean | undefined;
-}
-
-interface ButtonAreaProps extends SearchBarButtonProps {
+interface ButtonAreaProps extends RoundButtonProps {
   divide?: boolean;
   xs?: number;
   ariaLabel?: string;
