+++ frontend/src/components/Header/SearchBar/SelectItemArea/SelectItemTemplate/SelectItemTemplate.tsx
@@ -1,6 +1,10 @@
-import { Box, GridProps, Popover, PopoverProps } from "@mui/material";
+import { Box, Popover, PopoverProps } from "@mui/material";
 
-import { GridItem, modalStyle } from "./SelectItemTemplate.style";
+import {
+  GridItem,
+  modalStyle,
+  SelectItemTemplateProps,
+} from "./SelectItemTemplate.style";
 
 const SelectItemTemplate = ({
   divide,
@@ -30,7 +34,3 @@ const ModalTemplate = ({ ...props }: PopoverProps): JSX.Element => {
 };
 
 export { SelectItemTemplate, ModalTemplate };
-
-export interface SelectItemTemplateProps extends GridProps {
-  divide?: boolean;
-}
